organize filesystem stuff
Move all of our filesystem utility functions and filesystem/operating
system abstraction code under common/fs.

Get rid of WriteFileAtomic while I'm there, since it was only used
in one place, and AtomicFileWriter is better anyway.

fixes #3
