+++ common/fs/tempfile_test.go
@@ -13,7 +13,7 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-package objectserver
+package fs
 
 import (
 	"io/ioutil"
@@ -22,7 +22,6 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
-	"github.com/troubling/hummingbird/common"
 )
 
 func TestTempFile(t *testing.T) {
@@ -33,7 +32,7 @@ func TestTempFile(t *testing.T) {
 	require.Nil(t, err)
 	f.Write([]byte("some crap"))
 	require.Nil(t, f.Save(filepath.Join(dir, "somefile")))
-	require.True(t, common.Exists(filepath.Join(dir, "somefile")))
+	require.True(t, Exists(filepath.Join(dir, "somefile")))
 }
 
 func TestTempFileDirRemoved(t *testing.T) {
@@ -48,7 +47,7 @@ func TestTempFileDirRemoved(t *testing.T) {
 	f.Write([]byte("some crap"))
 	os.RemoveAll(dir)
 	require.Nil(t, f.Save(filepath.Join(dir, "somefile")))
-	require.True(t, common.Exists(filepath.Join(dir, "somefile")))
+	require.True(t, Exists(filepath.Join(dir, "somefile")))
 }
 
 func TestTempFileReplace(t *testing.T) {
@@ -60,7 +59,7 @@ func TestTempFileReplace(t *testing.T) {
 	require.Nil(t, err)
 	f.Write([]byte("some crap"))
 	require.Nil(t, f.Save(filepath.Join(dir, "somefile")))
-	require.True(t, common.Exists(filepath.Join(dir, "somefile")))
+	require.True(t, Exists(filepath.Join(dir, "somefile")))
 	data, err := ioutil.ReadFile(filepath.Join(dir, "somefile"))
 	require.Nil(t, err)
 	require.Equal(t, []byte("some crap"), data)
