+++ common/fs/xattr.go
@@ -13,7 +13,7 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-package xattr
+package fs
 
 import "errors"
