+++ objectserver/main.go
@@ -33,6 +33,7 @@ import (
 	"github.com/justinas/alice"
 	"github.com/troubling/hummingbird/common"
 	"github.com/troubling/hummingbird/common/conf"
+	"github.com/troubling/hummingbird/common/fs"
 	"github.com/troubling/hummingbird/common/srv"
 	"github.com/troubling/hummingbird/middleware"
 )
@@ -433,7 +434,7 @@ func (server *ObjectServer) AcquireDevice(next http.Handler) http.Handler {
 		if device, ok := vars["device"]; ok && device != "" {
 			devicePath := filepath.Join(server.driveRoot, device)
 			if server.checkMounts {
-				if mounted, err := common.IsMount(devicePath); err != nil || mounted != true {
+				if mounted, err := fs.IsMount(devicePath); err != nil || mounted != true {
 					vars["Method"] = request.Method
 					srv.CustomErrorResponse(writer, 507, vars)
 					return
@@ -468,7 +469,7 @@ func (server *ObjectServer) updateDeviceLocks(seconds int64) {
 		time.Sleep(reloadTime)
 		for _, key := range server.diskInUse.Keys() {
 			lockPath := filepath.Join(server.driveRoot, key, "lock_device")
-			if common.Exists(lockPath) {
+			if fs.Exists(lockPath) {
 				server.diskInUse.Lock(key)
 			} else {
 				server.diskInUse.Unlock(key)
