+++ objectserver/update_test.go
@@ -25,7 +25,7 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/require"
-	"github.com/troubling/hummingbird/common"
+	"github.com/troubling/hummingbird/common/fs"
 	"github.com/troubling/hummingbird/common/pickle"
 	"github.com/troubling/hummingbird/common/srv"
 )
@@ -99,7 +99,7 @@ func TestUpdateDeleteAt(t *testing.T) {
 	cs.Close()
 	server.updateDeleteAt(req, deleteAtStr, vars, &dl)
 	expectedFile := filepath.Join(ts.root, "sda", "async_pending", "8fc", "02cc012fe572f27e455edbea32da78fc-12345.6789")
-	require.True(t, common.Exists(expectedFile))
+	require.True(t, fs.Exists(expectedFile))
 	data, err := ioutil.ReadFile(expectedFile)
 	require.Nil(t, err)
 	a, err := pickle.PickleLoads(data)
@@ -127,7 +127,7 @@ func TestUpdateDeleteAtNoHeaders(t *testing.T) {
 	deleteAtStr := "1434707411"
 	server.updateDeleteAt(req, deleteAtStr, vars, &DummyLogger{})
 	expectedFile := filepath.Join(ts.root, "sda", "async_pending", "8fc", "02cc012fe572f27e455edbea32da78fc-12345.6789")
-	require.True(t, common.Exists(expectedFile))
+	require.True(t, fs.Exists(expectedFile))
 	data, err := ioutil.ReadFile(expectedFile)
 	require.Nil(t, err)
 	a, err := pickle.PickleLoads(data)
@@ -181,7 +181,7 @@ func TestUpdateContainer(t *testing.T) {
 	cs.Close()
 	server.updateContainer(metadata, req, vars, &dl)
 	expectedFile := filepath.Join(ts.root, "sda", "async_pending", "099", "2f714cd91b0e5d803cde2012b01d7099-12345.6789")
-	require.True(t, common.Exists(expectedFile))
+	require.True(t, fs.Exists(expectedFile))
 	data, err := ioutil.ReadFile(expectedFile)
 	require.Nil(t, err)
 	a, err := pickle.PickleLoads(data)
@@ -225,5 +225,5 @@ func TestUpdateContainerNoHeaders(t *testing.T) {
 	cs.Close()
 	server.updateContainer(metadata, req, vars, &dl)
 	expectedFile := filepath.Join(ts.root, "sda", "async_pending", "099", "2f714cd91b0e5d803cde2012b01d7099-12345.6789")
-	require.False(t, common.Exists(expectedFile))
+	require.False(t, fs.Exists(expectedFile))
 }
